refactored Logger
started unit testing

+++ RentItServer/RentItServer/Utilities/Logger.cs
@@ -22,9 +22,10 @@ namespace RentItServer.Utilities
 
         public Logger(string absolutePath)
         {
+            _absolutePath = absolutePath;
             String directory = absolutePath.Substring(0, absolutePath.LastIndexOf(Path.DirectorySeparatorChar) + 1);
             Directory.CreateDirectory(directory);
-            if (File.Exists(absolutePath))
+            if (File.Exists(absolutePath) == false)
             {
                 File.Create(absolutePath);
             }
@@ -47,22 +48,22 @@ namespace RentItServer.Utilities
         {
             String directory = absolutePath.Substring(0, absolutePath.LastIndexOf(Path.DirectorySeparatorChar) + 1);
             Directory.CreateDirectory(directory);
-            //if (File.Exists(absolutePath))
-            //{
-            File.Create(absolutePath);
-            //}
+            if (File.Exists(absolutePath) == false)
+            {
+                File.Create(absolutePath);
+            }
             this._absolutePath = absolutePath;
             handler += AddEntry;
             // Logging thread. Used in order to support asyncrhonous writing of entries
-            //new Task(() =>
-            //{
-            //    string logEntry;
-            //    while (true)
-            //    {
-            //        logEntry = _taskCollection.Take();
-            //        File.AppendAllText(absolutePath, logEntry);
-            //    }
-            //}).Start();
+            new Task(() =>
+            {
+                string logEntry;
+                while (true)
+                {
+                    logEntry = _taskCollection.Take();
+                    File.AppendAllText(absolutePath, logEntry);
+                }
+            }).Start();
         }
 
         /// <summary>
@@ -82,7 +83,7 @@ namespace RentItServer.Utilities
             {
                 string timeStamp = "[" + DateTime.UtcNow.ToString(CultureInfo.InvariantCulture) + "] ";
                 //File.AppendAllText(_absolutePath, timeStamp + args.Entry + Environment.NewLine); 
-                //_taskCollection.Add(timeStamp + args.Entry + Environment.NewLine);
+                _taskCollection.Add(timeStamp + args.Entry + Environment.NewLine);
             }
         }
     }
